remove tests unrelated to features in the application interface
