--- tests/test_workflow_parser.py
@@ -1,259 +1,0 @@
-from __future__ import annotations
-
-from pathlib import Path
-
-import pytest
-
-from myteam.workflow.parser import load_workflow
-
-
-def test_load_workflow_accepts_documented_workflow_shape(tmp_path: Path):
-    workflow_file = tmp_path / "haikus.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  prompt: I need three haikus. Please write them for me.\n"
-        "  output:\n"
-        "    haiku_dogs: A haiku about dogs\n"
-        "    haiku_cats: A haiku about cats\n"
-        "    haiku_user_choice: A haiku about a topic provided by the user\n"
-        "step2:\n"
-        "  input: $step1.output\n"
-        "  prompt: |\n"
-        "    Review the provided haikus.\n"
-        "    Rank them in terms of which best captures the essence and style of haiku.\n"
-        "  output:\n"
-        "    best_haiku:\n"
-        "      haiku: the haiku text\n"
-        "      reason: why this haiku was chosen over the others\n",
-        encoding="utf-8",
-    )
-
-    workflow = load_workflow(workflow_file)
-
-    assert list(workflow) == ["step1", "step2"]
-    assert workflow["step1"]["prompt"] == "I need three haikus. Please write them for me."
-    assert workflow["step1"]["input"] is None
-    assert workflow["step2"]["input"] == "$step1.output"
-    assert workflow["step2"]["output"]["best_haiku"]["haiku"] == "the haiku text"
-
-
-def test_load_workflow_defaults_missing_output_to_empty_mapping(tmp_path: Path):
-    workflow_file = tmp_path / "minimal.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  prompt: hello\n",
-        encoding="utf-8",
-    )
-
-    workflow = load_workflow(workflow_file)
-
-    assert workflow["step1"]["output"] == {}
-
-
-def test_load_workflow_rejects_steps_wrapper(tmp_path: Path):
-    workflow_file = tmp_path / "wrapped.yaml"
-    workflow_file.write_text(
-        "steps:\n"
-        "  step1:\n"
-        "    prompt: hello\n"
-        "    output:\n"
-        "      message: hi\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match="Workflow step 'steps' has unsupported keys: step1."):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_rejects_unknown_step_keys(tmp_path: Path):
-    workflow_file = tmp_path / "extra-key.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n"
-        "  timeout: 30\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match="unsupported keys: timeout"):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_rejects_non_identifier_keys_anywhere_in_structure(tmp_path: Path):
-    workflow_file = tmp_path / "bad-keys.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  prompt: hello\n"
-        "  input:\n"
-        "    bad-key: value\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match="non-identifier key: 'bad-key'"):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_rejects_unknown_agents(tmp_path: Path):
-    workflow_file = tmp_path / "agent.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  agent: mystery\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match="Unknown workflow agent: mystery"):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_accepts_model_and_extra_args(tmp_path: Path):
-    workflow_file = tmp_path / "agent-args.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  agent: codex\n"
-        "  interactive: false\n"
-        "  model: gpt-5.4\n"
-        "  extra_args:\n"
-        "    - --exec\n"
-        "    - pytest -q\n"
-        "  session_id: thread-123\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    workflow = load_workflow(workflow_file)
-
-    assert workflow["step1"]["model"] == "gpt-5.4"
-    assert workflow["step1"]["extra_args"] == ["--exec", "pytest -q"]
-    assert workflow["step1"]["interactive"] is False
-    assert workflow["step1"]["session_id"] == "thread-123"
-
-
-def test_load_workflow_rejects_invalid_model(tmp_path: Path):
-    workflow_file = tmp_path / "bad-model.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  model: ''\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match="field 'model' must be a non-empty string"):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_rejects_invalid_extra_args(tmp_path: Path):
-    workflow_file = tmp_path / "bad-extra-args.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  extra_args:\n"
-        "    - --exec\n"
-        "    - 3\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match=r"field 'extra_args\[1\]' must be a string"):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_rejects_invalid_interactive(tmp_path: Path):
-    workflow_file = tmp_path / "bad-interactive.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  interactive: 'yes'\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match="field 'interactive' must be a boolean"):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_rejects_fork_without_session_id(tmp_path: Path):
-    workflow_file = tmp_path / "bad-session-mode.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  fork: true\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match="field 'session_id' is required when 'fork' is true"):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_rejects_invalid_fork(tmp_path: Path):
-    workflow_file = tmp_path / "bad-fork-session.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  session_id: thread-123\n"
-        "  fork: 'yes'\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match="field 'fork' must be a boolean"):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_rejects_invalid_session_id(tmp_path: Path):
-    workflow_file = tmp_path / "bad-session.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  session_id: ''\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match="field 'session_id' must be a non-empty string"):
-        load_workflow(workflow_file)
-
-
-def test_load_workflow_normalizes_missing_input_to_null(tmp_path: Path):
-    workflow_file = tmp_path / "normalized.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    message: hi\n",
-        encoding="utf-8",
-    )
-
-    workflow = load_workflow(workflow_file)
-
-    assert workflow["step1"]["input"] is None
-
-
-def test_load_workflow_rejects_lists_in_output_templates(tmp_path: Path):
-    workflow_file = tmp_path / "lists.yaml"
-    workflow_file.write_text(
-        "step1:\n"
-        "  prompt: hello\n"
-        "  output:\n"
-        "    items:\n"
-        "      - title: item title\n",
-        encoding="utf-8",
-    )
-
-    with pytest.raises(ValueError, match=r"Workflow step 'step1'\.output\.items must not contain a list\."):
-        load_workflow(workflow_file)
